// apps/claw-frontend/src/enums/workspace-action-type.enum.ts
export enum WorkspaceActionType {
  CREATE_ISSUE = 'CREATE_ISSUE',
  CREATE_ISSUE_COMMENT = 'CREATE_ISSUE_COMMENT',
  CREATE_PR_DESCRIPTION = 'CREATE_PR_DESCRIPTION',
  CREATE_TICKET = 'CREATE_TICKET',
  ADD_TICKET_COMMENT = 'ADD_TICKET_COMMENT',
  SEND_SLACK_MESSAGE = 'SEND_SLACK_MESSAGE',
  SEND_EMAIL = 'SEND_EMAIL',
  REPLY_EMAIL = 'REPLY_EMAIL',
  CREATE_DRAFT = 'CREATE_DRAFT',
  UPDATE_JIRA_ISSUE = 'UPDATE_JIRA_ISSUE',
  COMMENT_JIRA = 'COMMENT_JIRA',
  COMMENT_PR = 'COMMENT_PR',
  APPROVE_PR = 'APPROVE_PR',
  ADD_PR_SUGGESTION = 'ADD_PR_SUGGESTION',
  SEND_SLACK = 'SEND_SLACK',
  REPLY_SLACK = 'REPLY_SLACK',
  EDIT_CONFLUENCE = 'EDIT_CONFLUENCE',
  CREATE_CONFLUENCE = 'CREATE_CONFLUENCE',
  UPLOAD_DRIVE = 'UPLOAD_DRIVE',
  MOVE_DRIVE = 'MOVE_DRIVE',
  CREATE_USER_STORY_FROM_FIGMA = 'CREATE_USER_STORY_FROM_FIGMA',
  CREATE_JIRA_FROM_FIGMA = 'CREATE_JIRA_FROM_FIGMA',
  POST_FIGMA_COMMENT = 'POST_FIGMA_COMMENT',
  CREATE_MR_COMMENT = 'CREATE_MR_COMMENT',
  APPROVE_MR = 'APPROVE_MR',
  CREATE_GITLAB_ISSUE = 'CREATE_GITLAB_ISSUE',
  COMMENT_GITLAB_ISSUE = 'COMMENT_GITLAB_ISSUE',
  UPDATE_MR_DESCRIPTION = 'UPDATE_MR_DESCRIPTION',
  ADD_MR_SUGGESTION = 'ADD_MR_SUGGESTION',
  ADD_MR_IMAGE_COMMENT = 'ADD_MR_IMAGE_COMMENT',
  CREATE_PR_COMMENT_BB = 'CREATE_PR_COMMENT_BB',
  APPROVE_PR_BB = 'APPROVE_PR_BB',
  CREATE_BITBUCKET_ISSUE = 'CREATE_BITBUCKET_ISSUE',
  UPLOAD_ONEDRIVE = 'UPLOAD_ONEDRIVE',
  MOVE_ONEDRIVE = 'MOVE_ONEDRIVE',
  UPLOAD_SHAREPOINT = 'UPLOAD_SHAREPOINT',
  CREATE_SHAREPOINT_LIST_ITEM = 'CREATE_SHAREPOINT_LIST_ITEM',
  UPDATE_SHAREPOINT_LIST_ITEM = 'UPDATE_SHAREPOINT_LIST_ITEM',
  CREATE_CLICKUP_TASK = 'CREATE_CLICKUP_TASK',
  UPDATE_CLICKUP_TASK = 'UPDATE_CLICKUP_TASK',
  COMMENT_CLICKUP_TASK = 'COMMENT_CLICKUP_TASK',
}